backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.0
chromadb==0.4.24
openai==1.14.0
sentence-transformers==2.5.1
python-dotenv==1.0.1
numpy==1.26.4
httpx==0.27.0
reportlab==3.6.12
loguru==0.7.2
python-dateutil==2.8.2
